+++ infrastructure/database/models/user_tokens.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, ForeignKey
+from sqlalchemy import Integer, String, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .base import BaseModel
@@ -7,10 +7,10 @@ from .base import BaseModel
 class UserTokens(BaseModel):
     __tablename__ = "user_tokens"
 
-    user_id: Mapped[int] = mapped_column(Integer)
+    user_id: Mapped[int] = mapped_column(BigInteger)
     token: Mapped[str] = mapped_column(String, unique=True)
     order_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("subscription_orders.query_id"), unique=True
+        BigInteger, ForeignKey("subscription_orders.query_id"), unique=True
     )
     days: Mapped[int] = mapped_column(Integer)
     date: Mapped[int] = mapped_column(Integer)
